+++ pureml/cli/params.py
@@ -11,22 +11,24 @@ import json
 import typing
 from urllib.parse import urljoin
 
-from . import get_token, get_project_id, BASE_URL, PATH_PURE_DIR
+from . import get_token, get_project_id, get_org_id, BASE_URL, PATH_PURE_DIR
 
 
 app = typer.Typer()
 
 
 @app.command()
-def add(params, model_name: str=None) -> str:
+def add(params, model_name: str, model_version:str) -> str:
     '''`add()` takes a dictionary of parameters and a model name as input and returns a string
     
     Parameters
     ----------
-    params
+    params : dict
         a dictionary of parameters
     model_name : str
         The name of the model you want to add parameters to.
+    model_version: str
+        The version of the model
     
     Returns
     -------
@@ -35,10 +37,10 @@ def add(params, model_name: str=None) -> str:
     '''
 
     user_token = get_token()
+    org_id = get_org_id()
     project_id = get_project_id()
     
-
-    url_path_1 = 'model/{}/{}/params/add'.format(project_id, model_name)
+    url_path_1 = '{}/project/{}/model/{}/params/add'.format(org_id, project_id, model_name)
     url = urljoin(BASE_URL, url_path_1)
 
 
@@ -56,15 +58,17 @@ def add(params, model_name: str=None) -> str:
     if response.status_code == 200:
         print(f"[bold green]Params have been registered!")
 
-        return response.text
     else:
         print(f"[bold red]Params have not been registered!")
-        return
+
+    
+    return response.text
+        
 
 
 @app.command()
-def fetch(model_name: str, param:str='') -> str:
-    '''`fetch(model_name: str, param:str='') -> str`
+def fetch(model_name: str, model_version:str, param:str='') -> str:
+    '''
     
     This function fetches the parameters of a model
     
@@ -72,6 +76,8 @@ def fetch(model_name: str, param:str='') -> str:
     ----------
     model_name : str
         The name of the model you want to fetch the parameters for.
+    model_version: str
+        The version of the model
     param : str
         The name of the parameter to fetch. If not specified, all parameters are returned.
     
@@ -81,10 +87,11 @@ def fetch(model_name: str, param:str='') -> str:
     
     '''
     user_token = get_token()
+    org_id = get_org_id()
     project_id = get_project_id()
     
 
-    url_path_1 = 'model/{}/{}/params/{}'.format(project_id, model_name, param)
+    url_path_1 = '{}/project/{}/model/{}/params/{}'.format(org_id, project_id, model_name, param)
     url = urljoin(BASE_URL, url_path_1)
 
 
@@ -134,7 +141,7 @@ def fetch(model_name: str, param:str='') -> str:
 
 
 @app.command()
-def delete(model_name:str, param:str) -> str:
+def delete(param:str, model_name:str, model_version:str) -> str:
     '''This function deletes a parameter from a model
     
     Parameters
@@ -143,13 +150,16 @@ def delete(model_name:str, param:str) -> str:
         The name of the model you want to delete the parameter from.
     param : str
         The name of the parameter to delete.
+    model_version: str
+        The version of the model
     
     '''
     user_token = get_token()
+    org_id = get_org_id()
     project_id = get_project_id()
     
 
-    url_path_1 = 'model/{}/{}/params/{}/delete'.format(project_id, model_name, param)
+    url_path_1 = '{}/project/{}/model/{}/params/{}/delete'.format(org_id,project_id, model_name, param)
     url = urljoin(BASE_URL, url_path_1)